chore: add -l/--link-libs to `cargo run -- run -l` to link libraries instead of copying those (only linux)
